Separated out SkyBox from CookBook. Added Skybox as a component class in QuakeEngine. Currently unsure as to the correctness of the approach but it works. To be reviewed.
