+++ src/TK-Quake.Engine/TK-Quake.Engine.csproj
@@ -62,6 +62,7 @@
     <Compile Include="Infrastructure\Components\BobComponent.cs" />
     <Compile Include="Infrastructure\Components\IComponent.cs" />
     <Compile Include="Infrastructure\Components\RotateOnUpdateComponent.cs" />
+    <Compile Include="Infrastructure\Components\SkyBoxComponent.cs" />
     <Compile Include="Infrastructure\Entities\IEntity.cs" />
     <Compile Include="Infrastructure\Entities\Entity.cs" />
     <Compile Include="Infrastructure\Entities\RenderableEntity.cs" />
